refractoring of actions, better remote support impl in progress

+++ actions/Cell.js
@@ -18,21 +18,50 @@ module.exports = organic.Organel.extend(function Cell(plasma, config){
         "npm install; "
       var child = shelljs.exec("ssh "+c.remote+' "'+cmd+'"', {async: true});
       child.on('exit', function(code){
-        console.log('done with code ', code);
+        if(callback) callback({data: code});
       });
     } else {
       shelljs.mkdir('-p', c.target);
       shelljs.exec("git clone "+c.source+" "+c.target);
       shelljs.cd(c.target);
       shelljs.exec("npm install");
+      if(callback) callback({data: 0});
     }
   },
   "upgrade": function(c, sender, callback){
-    process.kill(c.target, "SIGUSR1");
-    if(callback) callback();
+    if(c.remote) {
+      var cmd = ". ~/.nvm/nvm.sh; nvm use "+process.version+"; "+
+        "cd "+c.target+"; git pull; "+
+        "npm install; "
+      var child = shelljs.exec("ssh "+c.remote+' "'+cmd+'"', {async: true});
+      child.on('exit', function(code){
+        if(callback) callback({data: code});
+      });
+    } else {
+      process.kill(c.target, "SIGUSR1");
+      if(callback) callback();
+    }
   },
   "restart": function(c, sender, callback){
     process.kill(c.target, "SIGUSR2");
     if(callback) callback();
+  },
+  "start":  function(c, sender, callback){
+    if(c.remote) {
+      var cmd = ". ~/.nvm/nvm.sh; nvm use "+process.version+"; "+
+        "cd "+path.dirname(c.target)+"; git pull; "+
+        "npm install; "+
+        "angel Cell start "+c.target;
+      var child = shelljs.exec("ssh "+c.remote+' "'+cmd+'"', {async: true});
+      child.on('exit', function(code){
+        if(callback) callback({data: code});
+      });
+    } else {
+      this.emit({
+        type: "Tissue",
+        action: "start",
+        target: c.target
+      }, callback);
+    }
   }
 });
